+++ src/main/python/plotlyst/view/widget/character/__init__.py
@@ -1,0 +1,1 @@
+from .comp import CharacterComparisonWidget, LayoutType
