+++ src/main/python/plotlyst/view/widget/characters.py
@@ -70,9 +70,8 @@ from src.main.python.plotlyst.view.generated.scene_conflict_intensity_ui import 
 from src.main.python.plotlyst.view.generated.scene_dstribution_widget_ui import Ui_CharactersScenesDistributionWidget
 from src.main.python.plotlyst.view.generated.scene_goal_stakes_ui import Ui_GoalReferenceStakesEditor
 from src.main.python.plotlyst.view.icons import avatars, IconRegistry, set_avatar
-from src.main.python.plotlyst.view.widget.big_five import BigFiveChart, dimension_from
 from src.main.python.plotlyst.view.widget.button import SelectionItemPushButton
-from src.main.python.plotlyst.view.widget.display import IconText, Icon, RoleIcon, ChartView
+from src.main.python.plotlyst.view.widget.display import IconText, Icon
 from src.main.python.plotlyst.view.widget.input import DocumentTextEditor
 from src.main.python.plotlyst.view.widget.labels import ConflictLabel, CharacterLabel, CharacterGoalLabel
 from src.main.python.plotlyst.view.widget.progress import CircularProgressBar, ProgressTooltipMode, \
@@ -1939,44 +1938,3 @@ class CharacterTopicsEditor(QWidget):
                 self.setIcon(IconRegistry.from_name(topic.icon, topic.icon_color))
             self.setText(topic.text)
             self.setToolTip(topic.description)
-
-
-class CharacterOverviewWidget(QWidget):
-    def __init__(self, character: Character, parent=None):
-        super().__init__(parent)
-        self._character = character
-
-        self._avatar = QLabel(self)
-        set_avatar(self._avatar, self._character, size=118)
-        self._roleIcon = RoleIcon(self)
-        if self._character.role:
-            self._roleIcon.setRole(self._character.role, showText=True)
-
-        vbox(self, 0)
-        self.layout().addWidget(self._avatar, alignment=Qt.AlignmentFlag.AlignCenter)
-        self.layout().addWidget(self._roleIcon, alignment=Qt.AlignmentFlag.AlignCenter)
-        self.layout().addWidget(line())
-        self._bigFive = BigFiveChart()
-        self._bigFive.setTitle('')
-        for bf, values in character.big_five.items():
-            self._bigFive.refreshDimension(dimension_from(bf), values)
-        self._bigFiveChartView = ChartView(self)
-        self._bigFiveChartView.setChart(self._bigFive)
-
-        self.layout().addWidget(self._bigFiveChartView)
-
-
-class CharacterComparisonWidget(QWidget):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._characters: Dict[Character, CharacterOverviewWidget] = {}
-        hbox(self, spacing=0)
-
-    def updateCharacter(self, character: Character, enabled: bool):
-        if enabled:
-            wdg = CharacterOverviewWidget(character)
-            self._characters[character] = wdg
-            self.layout().addWidget(wdg)
-        else:
-            wdg = self._characters.pop(character)
-            self.layout().removeWidget(wdg)
